emotion_gui/emotion_gui/requirements.txt
opencv-python
tensorflow
keras
numpy
pillow